# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgriSense API",
    "description": "Backend API for AgriSense \u2014 SIH PS 26132",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/auth/otp/request": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Request Otp",
        "description": "Request an OTP for a phone number.",
        "operationId": "request_otp_api_v1_auth_otp_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestRes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/otp/verify": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Verify Otp",
        "description": "Verify OTP and login an existing user.",
        "operationId": "verify_otp_api_v1_auth_otp_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register",
        "description": "Register a new user after verifying OTP.",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get Me",
        "description": "Get the current authenticated user's profile.",
        "operationId": "get_me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/locale": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Update Locale",
        "description": "Update the user's preferred locale.",
        "operationId": "update_locale_api_v1_auth_locale_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocaleUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/lots": {
      "post": {
        "tags": [
          "Lots"
        ],
        "summary": "Create Lot",
        "description": "Create a new lot. Auto-calls Price Engine for price band.",
        "operationId": "create_lot_api_v1_lots_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LotCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LotOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Lots"
        ],
        "summary": "List Lots",
        "description": "List lots owned by the current farmer, optionally filtered by status.",
        "operationId": "list_lots_api_v1_lots_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LotOut"
                  },
                  "title": "Response List Lots Api V1 Lots Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lots/{lot_id}": {
      "get": {
        "tags": [
          "Lots"
        ],
        "summary": "Get Lot",
        "description": "Get a single lot. Checks ownership or buyer-has-offer access.",
        "operationId": "get_lot_api_v1_lots__lot_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LotOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lots/{lot_id}/assay": {
      "post": {
        "tags": [
          "Lots"
        ],
        "summary": "Submit Self Assay",
        "description": "Submit self-assay quality answers for a lot. Returns computed grade.",
        "operationId": "submit_self_assay_api_v1_lots__lot_id__assay_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lot Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelfAssayRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelfAssayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lots/{lot_id}/price-suggestion": {
      "get": {
        "tags": [
          "Lots"
        ],
        "summary": "Get Price Suggestion",
        "description": "Get price band + SELL/HOLD/NO_ADVICE recommendation for a lot.",
        "operationId": "get_price_suggestion_api_v1_lots__lot_id__price_suggestion_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceSuggestionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lots/{lot_id}/matches": {
      "get": {
        "tags": [
          "Lots"
        ],
        "summary": "Get Lot Matches",
        "description": "Find matching buyer demands for a lot.",
        "operationId": "get_lot_matches_api_v1_lots__lot_id__matches_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchesRes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/lots/{lot_id}/photo": {
      "post": {
        "tags": [
          "Lots"
        ],
        "summary": "Upload Lot Photo",
        "description": "Upload lot photo, strip EXIF GPS, and save.",
        "operationId": "upload_lot_photo_api_v1_lots__lot_id__photo_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "lot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Lot Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_lot_photo_api_v1_lots__lot_id__photo_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "user",
          "token"
        ],
        "title": "AuthResponse"
      },
...
      "Body_upload_lot_photo_api_v1_lots__lot_id__photo_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_lot_photo_api_v1_lots__lot_id__photo_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LocaleUpdateRequest": {
        "properties": {
          "locale": {
            "type": "string",
            "title": "Locale"
          }
        },
        "type": "object",
        "required": [
          "locale"
        ],
        "title": "LocaleUpdateRequest"
      },
...
      "LotCreate": {
        "properties": {
          "commodity_id": {
            "type": "string",
            "title": "Commodity Id"
          },
          "qty_kg": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Qty Kg"
          },
          "expected_price_paise": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Expected Price Paise"
          },
          "market_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Id"
          },
          "warehouse_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warehouse Id"
          },
          "self_assay": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Self Assay"
          }
        },
        "type": "object",
        "required": [
          "commodity_id",
          "qty_kg",
          "expected_price_paise"
        ],
        "title": "LotCreate"
      },
      "LotOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "commodity_id": {
            "type": "string",
            "title": "Commodity Id"
          },
          "qty_kg": {
            "type": "integer",
            "title": "Qty Kg"
          },
          "expected_price_paise": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Price Paise"
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "LISTED",
              "IN_AUCTION",
              "SOLD",
              "CANCELLED",
              "COMPLETED"
            ],
            "title": "Status"
          },
          "market_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Id"
          },
          "warehouse_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warehouse Id"
          },
          "grade": {
            "type": "string",
            "enum": [
              "UNGRADED",
              "A",
              "B",
              "C",
              "REJECTED"
            ],
            "title": "Grade"
          },
          "photo_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Path"
          },
          "farmer_id": {
            "type": "string",
            "title": "Farmer Id"
          },
          "harvest_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Harvest Date"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "commodity_id",
          "qty_kg",
          "status",
          "grade",
          "farmer_id"
        ],
        "title": "LotOut"
      },
...
      "MatchesRes": {
        "properties": {
          "matches": {
            "items": {
              "$ref": "#/components/schemas/MatchRow"
            },
            "type": "array",
            "title": "Matches"
          }
        },
        "type": "object",
        "required": [
          "matches"
        ],
        "title": "MatchesRes"
      },
      "MeResponse": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserDetailResponse"
          }
        },
        "type": "object",
        "required": [
          "user"
        ],
        "title": "MeResponse"
      },
...
      "OtpRequestReq": {
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 15,
            "minLength": 10,
            "title": "Phone"
          }
        },
        "type": "object",
        "required": [
          "phone"
        ],
        "title": "OtpRequestReq"
      },
      "OtpRequestRes": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "expires_in_s": {
            "type": "integer",
            "title": "Expires In S"
          },
          "dev_otp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dev Otp"
          }
        },
        "type": "object",
        "required": [
          "expires_in_s"
        ],
        "title": "OtpRequestRes"
      },
      "OtpVerifyReq": {
        "properties": {
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "code": {
            "type": "string",
            "maxLength": 6,
            "minLength": 6,
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "code"
        ],
        "title": "OtpVerifyReq"
      },
...
      "PriceSuggestionResponse": {
        "properties": {
          "lot_id": {
            "type": "string",
            "title": "Lot Id"
          },
          "price_band": {
            "$ref": "#/components/schemas/PriceBand"
          },
          "sale_window": {
            "$ref": "#/components/schemas/SaleWindowInfo"
          }
        },
        "type": "object",
        "required": [
          "lot_id",
          "price_band",
          "sale_window"
        ],
        "title": "PriceSuggestionResponse"
      },
      "RegisterRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "phone": {
            "type": "string",
            "maxLength": 15,
            "minLength": 10,
            "title": "Phone"
          },
          "code": {
            "type": "string",
            "maxLength": 6,
            "minLength": 6,
            "title": "Code"
          },
          "role": {
            "type": "string",
            "pattern": "^(FARMER|BUYER|FPO_ADMIN)$",
            "title": "Role"
          },
          "locale": {
            "type": "string",
            "title": "Locale",
            "default": "mr"
          },
          "district_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District Id"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lng"
          },
          "company_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Name"
          }
        },
        "type": "object",
        "required": [
          "name",
          "phone",
          "code",
          "role"
        ],
        "title": "RegisterRequest"
      },
...
      "SelfAssayRequest": {
        "properties": {
          "size_uniform": {
            "type": "integer",
            "enum": [
              1,
              2,
              3
            ],
            "title": "Size Uniform"
          },
          "colour_uniform": {
            "type": "integer",
            "enum": [
              1,
              2,
              3
            ],
            "title": "Colour Uniform"
          },
          "damage_pct": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Damage Pct"
          },
          "sprouting": {
            "type": "integer",
            "enum": [
              1,
              2,
              3
            ],
            "title": "Sprouting"
          },
          "moisture_feel": {
            "type": "integer",
            "enum": [
              1,
              2,
              3
            ],
            "title": "Moisture Feel"
          },
          "foreign_matter": {
            "type": "integer",
            "enum": [
              1,
              2,
              3
            ],
            "title": "Foreign Matter"
          }
        },
        "type": "object",
        "required": [
          "size_uniform",
          "colour_uniform",
          "damage_pct",
          "sprouting",
          "moisture_feel",
          "foreign_matter"
        ],
        "title": "SelfAssayRequest"
      },
      "SelfAssayResponse": {
        "properties": {
          "lot_id": {
            "type": "string",
            "title": "Lot Id"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "grade": {
            "type": "string",
            "title": "Grade"
          },
          "weakest_dimension": {
            "type": "string",
            "title": "Weakest Dimension"
          },
          "tip_mr": {
            "type": "string",
            "title": "Tip Mr"
          },
          "tip_en": {
            "type": "string",
            "title": "Tip En"
          },
          "self_assay_answers": {
            "additionalProperties": true,
            "type": "object",
            "title": "Self Assay Answers"
          }
        },
        "type": "object",
        "required": [
          "lot_id",
          "score",
          "grade",
          "weakest_dimension",
          "tip_mr",
          "tip_en",
          "self_assay_answers"
        ],
        "title": "SelfAssayResponse"
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "locale": {
            "type": "string",
            "title": "Locale"
          },
          "district_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "phone",
          "role",
          "locale"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}